refactor(nms): Refactor nms top level directories
