+++ nms/docker-compose.yml
@@ -31,7 +31,6 @@ services:
       - ./scripts:/usr/src/scripts
       - ./server:/usr/src/server
       - ./shared:/usr/src/shared
-      - ./views:/usr/src/views
     depends_on:
       - postgres
     networks:
