+++ nms/server/alerts/lteAlerts.js
@@ -14,7 +14,7 @@
  * @format
  */
 
-import type {prom_alert_config} from '../generated/MagmaAPIBindings';
+import type {prom_alert_config} from '../../generated/MagmaAPIBindings';
 
 export default function getLteAlerts(
   networkID: string,
