+++ nms/server/alerts/sync.js
@@ -14,18 +14,18 @@
  * @format
  */
 
-import MagmaV1API from '../server/magma/index';
+import MagmaV1API from '../../server/magma/index';
 import getCwfAlerts from './cwfAlerts';
 import getFegAlerts from './fegAlerts';
 import getLteAlerts from './lteAlerts';
 
-import {CWF, FEG, FEG_LTE, LTE} from '../shared/types/network';
+import {CWF, FEG, FEG_LTE, LTE} from '../../shared/types/network';
 import type {ExpressResponse} from 'express';
-import type {FBCNMSRequest} from '../server/auth/access';
+import type {FBCNMSRequest} from '../../server/auth/access';
 import type {
   network_type,
   prom_alert_config,
-} from '../generated/MagmaAPIBindings';
+} from '../../generated/MagmaAPIBindings';
 
 async function syncAlertsForNetwork(
   networkID: string,
