+++ nms/server/views/host.pug
